Move windows specific code to its own file
