+++ Lib/test/libregrtest/main.py
@@ -18,8 +18,7 @@ from test.libregrtest.runtest import (
     INTERRUPTED, CHILD_ERROR, TEST_DID_NOT_RUN,
     PROGRESS_MIN_TIME, format_test_result)
 from test.libregrtest.setup import setup_tests
-from test.libregrtest.utils import (
-    removepy, count, format_duration, printlist, WindowsLoadTracker)
+from test.libregrtest.utils import removepy, count, format_duration, printlist
 from test import support
 try:
     import gc
@@ -595,6 +594,8 @@ class Regrtest:
                 return os.getloadavg()[0]
             self.getloadavg = getloadavg_1m
         elif sys.platform == 'win32' and (self.ns.slaveargs is None):
+            from test.libregrtest.win_utils import WindowsLoadTracker
+
             load_tracker = WindowsLoadTracker()
             self.getloadavg = load_tracker.getloadavg
